requirements.txt
transformers>=4.39.0
torch>=2.1.0
torchvision>=0.16.0
pillow>=10.0.0
accelerate>=0.25.0
bitsandbytes>=0.43.0
datasets>=2.14.0
soundfile>=0.12.1
pynput>=1.7.6
mss>=9.0.1
sounddevice>=0.4.6